apiObjectModle

+++ platform_api/src/main/java/com/gongxc/platform_api/controller/CaseController.java
@@ -1,23 +1,37 @@
 package com.gongxc.platform_api.controller;
 
 import com.gongxc.platform_api.common.ResponseResult;
+import com.gongxc.platform_api.entity.ApiObjectModel;
 import com.gongxc.platform_api.entity.Cases;
+import com.gongxc.platform_api.entity.ResponseData;
+import com.gongxc.platform_api.service.ApiService;
 import com.gongxc.platform_api.service.CaseService;
+import com.gongxc.platform_api.service.ResultService;
+import io.restassured.specification.RequestSpecification;
 import io.swagger.annotations.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
+import springfox.documentation.spring.web.json.Json;
 
 import java.util.List;
 import java.util.Optional;
 
+import static io.restassured.RestAssured.given;
+
 @RestController
 @RequestMapping(value = "/case",produces = "application/json")
 @Api(value="Cases",tags = {"测试平台"}, description = "接口测试平台")
-public class CaseControllor {
+public class CaseController {
 
     @Autowired
     private CaseService caseService;
+
+    @Autowired
+    private ResultService resultService;
+
+    @Autowired
+    private ApiService apiService;
 
     @GetMapping("/findById/{id}")
     @ApiOperation(value = "使用id查询用例")
@@ -31,7 +45,7 @@ public class CaseControllor {
     })
     public ResponseResult<Optional<Cases>> findcaseById(@PathVariable("id") Long id){
 
-        Optional<Cases> cases= caseService.findUserById(id);
+        Optional<Cases> cases= caseService.findCaseById(id);
         return ResponseResult.successWithData(cases);
     }
 
@@ -54,9 +68,15 @@ public class CaseControllor {
             @ApiResponse(code = 500, message = "服务器内部错误")
     })
     public ResponseResult<List<Cases>> deleteById(Long id){
-        caseService.deleteById(id);
-        List<Cases> casess = caseService.findAll();
-        return ResponseResult.successWithData(casess);
+        Optional<Cases> cases = caseService.findCaseById(id);
+        if (cases.isPresent()){
+            caseService.deleteById(id);
+            List<Cases> casess = caseService.findAll();
+            return ResponseResult.successWithData(casess);
+        }else {
+            return ResponseResult.failWithCodeAndMsg(1000,"用例不存在不存在");
+        }
+
     }
 
     @PostMapping("/create")
@@ -96,8 +116,24 @@ public class CaseControllor {
     })
     public ResponseResult<Optional<Cases>> updatecases(Long id,String name){
         caseService.update(id,name);
-        Optional<Cases> cases1 = caseService.findUserById(id);
+        Optional<Cases> cases1 = caseService.findCaseById(id);
         return ResponseResult.successWithData(cases1);
+    }
+
+    @PostMapping("/run")
+    @ApiOperation(value = "执行用例")
+    @ApiImplicitParams({
+            @ApiImplicitParam(value = "ID",name = "id",dataType = "int",required = true,defaultValue = "1"),
+    })
+    @ApiResponses({
+            @ApiResponse(code = 401,message = "未授权"),
+            @ApiResponse(code = 404,message = "请求路径不存在"),
+            @ApiResponse(code = 500, message = "服务器内部错误")
+    })
+    public ResponseResult<Optional<Cases>> runCases(Long id){
+        caseService.run(id);
+        Optional<Cases> cases = caseService.findCaseById(id);
+        return ResponseResult.successWithData(cases);
     }
 
 }
